- rpwmrc: make sure stderr shows up in the output (?)
actually, don't remember why this was needed.  just ran a 'verbexec'
test and neither stdout nor stderr even show up at all, despite the name
"verbexec" so not sure why this was present, but it wouldn't be there if
it didn't have some purpose.  maybe will figure it out later...