+ etc: keymap: add linux console keymap from the old days
this was how it all started... one of the earliest things customized.
only minor modifications have been made since.  even the nice newlines
are original.

some day, maybe start using one of the fast kms fbcons for normal tmux,
and not use X for thatg.  however with no real browser options for
directfb this is unlikely.  (there's a firefox project, but it looks
like it might be dead...)