+ rpwmrc: use an X compositor, stops apps like chrome screen-tearing
there's tearing in chrome, etc when compositing is not used, so run a
persistent compositor

it seems that there's an additional buffering step done when compositing
(ie, to support a "composition" phase) so that all apps can input their
contribution to the compositor, which does a final draw at the end.  in
addition to its ostensible purpose, this seems to eliminate tears.
since KDE/Gnome etc have a compositor and are the ubiquitous Linux
desktop, this is probably the test environment used for Chrome and other
apps, so they just assume a compositor is present.  using 'compton'
seems to be a favorite solution of people that use other tiling window
managers like i3, and does seem to work well.  it also fixes drawing
glitches and redraw flashes every second when a ratpoison menu is
visible.  there's no apparent downside, although once in a while there's
a weird delay when the screen is black or green-black stippled before
the "real" composed screen image is drawn; this is only a few hundred ms
and happens only a couple times per day, if at all.  there's probably
some imperceptible draw latency also due to the additional buffering,
but we'll definitely take it instead of the horrible tearing that
occurs without it

note that xsetroot stopped working with the compositor running, so we
apparently have to use a composed version of setroot: hsetroot, also
included in this patch