+ rpwmrc: attempt to make browser use VA-API for hwaccel video
this makes the "hardware accelerated" flag turn green in brave://gpu,
but doesn't actually work, as shown in devtools on a youtube video, it's
actually using software decode.  it might be some combination of using
vulkan or not having the right driver settings or libraries.  played
with this for a couple days and was unable to get it to work.  since
video on my laptop is a very infrequent activity, this shouldn't affect
me much, and it's fine to just leave this disabled.

some future update might work, or might play more with it later, so
leaving the flag enabled for now.