+ xkeymap: switch to xkb from xmodmap, run from xinputs
running xmodmap several times on keyboard changes never worked, but
running with xkb replaces the entire keymap and seems to be effectively
idempotent.  this lets us now plug keyboard in and just run 'xinputs'
and everything "just works"

along the way we fixed the chrom not recognizing remapped backslash/vbar
key by letting the regular key retain its original function when in
Keyboard Group 2, and we've made shift-escape toggle this group.  still
lame that we have to switch the group just to use backslash in a chrome
textarea, but this is an old bug (since at least 2014) and unlikely to
be fixed soon: https://bugs.chromium.org/p/chromium/issues/detail?id=394552