+ guibrose: enable gpu rasterization in the spawn flags
"Chrome 37 introduced a GPU rasterizer. When enabled, some paint
workloads can go from 100ms/frame to 4-5ms/frame.  It doesn't run for
all content, but is enabled as a combination of viewport settings and a
device allow-list."

we add it here instead of chrome flags internally, so it's documentated.
probably all of them should be like that but we'll just use ones going
forward and maybe move some from config later (and, it's not clear that
all chrome://flags correspond to command line options)