/ rpwmrc: sleep before urxvtc to give Xresources a chance to load
lame consequence of ratpoison's 'exec' and 'verbexec' being asynchronous
forked commands that are only waited on by signal handler.  they don't
block so everything in the ratpoison init basically races.  should do
something about this but the best solution is for ratpoison to provide a
synchronous exec command and that requires maintaining our own forked
ratpoison.  maybe this is ok since it's not developed anymore really
anyways