- keys: M-t: run 'taskline' instead of time (displays time and more)
we want the current task and whether the clock is on, to be available
just as frequently as the date.  recently added .task/ 'taskline'
utility will do this: combines current fql, whether time clock is active
or not, and time+date, and has them all output at once.  so we'll use
it.

reason for the env set is, that's done in shell profile, and isn't in
the base ratpoison environment.  we could maybe add it, but let's see if
this is ever used in more than one place