+ zoomrc: add new rcfile for zoom's config, used from the firejail
the defaults that come with firejail don't work, in particular netlink
is needed for some reason, which seems bizarre for an app to need, as
this is for communicating with the kernel.  am a bit worried about this
but will have to research it later.

also we need alsa config to be shared into the container