- rpwmrc: use drm for vsync and and adjust glx params for u22
after u20 -> u22 upgrade, the screen was black until compton was killed.
something changed, fiddled with the flags.

it could be suboptimal because didn't exhast the full permutation space of
options but a fair number to get what seemed best

probably it needs to be retuned on each new kernel and/org userspace
drm/glx stuff.  who knows