+ asoundrc: store alsa rcfile, not xorg but no better place atm
if we end up getting several files like this that really aren't related
to xorg, we will have to either rename this repository to be more
generalized or have another repository for non-X dotfiles

it's still somewhat "related" to xorg, insofar as it would likely only
be used on a "full" console system that was running xorg on the console.
this is typically the only place where we use sound, ie on the laptop.

this is also the reason we're checking it into 'omnius' branch, because
alsa config file is obviously host-specific