+ jcsrc: tweak ratpoison config to suit sdorfehs
this will only work with all my patches 36, 38, 41, 42, 43

in that case it tries to be like the old ratpoison behavior with these
settings.  there's a couple optimizations like the window will remain on
top.  have not quite figured out everything yet but this makes me able
to move from ratpoison and is a good first cut