+ jcsrc: keys: enter M-m, then any key, to send M-key to window
have needed on somewhat rare occasion to use M-foo but not been able to
since we don't have mappings for that.  we are using M-stuff for
toplevel control of window manager, so those are really reserved for its
use.  instead we can use a prefix key sent beforehand, next key read
will have meta prefixed and will be injected into the current window's
key input queue.

example: using the two-key sequence "M-m a" will send the one-key
sequence "M-a" to the application