+ xrdb: st: usebold: disable bold weight, terminals have intensity
this is adjustment from my own patch 'usebold', its default keeps the st
default to do bold rendering, but this causes the fonts to be larger,
which is strange.  as long as this author can remember bold has only
meant high intensity because terminals were not capable of different
widths for the fonts.  also this causes unsightly gaps in the lines
which are no fun, we do not want that!