- xrdb: urxvt: lightGrey -> lightgrey, so it's lowercase
most things that match it appear to use insensitive so case doesn't
matter.  after trying urxvt with both cases, do not see a difference, it
must be insensitive.  confirmed with internet search most uses are case
insensitive so we will use that.  we plan to use them in upcoming
commits for st so we will want to use consistent naming case convention