+ guibrowse: wake up existing browser process group before spawn
often we forget that we paused it, run a new instance, chrome times out
waiting to contact it, and then starts a new session, thinking there was
a hard crash.  this can lose session data.

instead we'll just always try to resume chrome, whether it's paused or
not.  there's still probably a race between completion of resumption and
the new socket instance, but it surely waits for some time, it should
get queued.  once we send the kills they should get scheduled before our
process resumes.  will see if this ever breaks in practice