+ jcsrc: add taskwarrior/timewarrior mappings to rptaskline script
we have added the script in src/utilx@e604ca6 which wraps prompt input
for commands which need it and output of a new taskline in ratpoison
after the command has been run.  intended for use to start/stop/switch
tasks.  it also does the existing taskline function when called with no
arguments, so there are only additional features here