+ guibrowse: add profile selection flags -p/-w, default personal
we were passing it in from the window manager depending on what key was
pressed, but this means that we can only use the window manager, never
the shell, to start something in the work profile.  unless we spell the
whole thing out.  easier to just give a -p and -w option