+ jcsrc: keys: M-S: need a 'taskcont', not just 'timecont', so juggle
for all our taskwtools binds, we have used the paradigm M-<X> for 'task'
and M-<x> for 'time' variants.  for 'stop' though, we actually do not
have use for two versions because whether a task or pseudo-task in
time/, stop action is the same, so we only needed one key for that.
instead we overloaded 's' to act as both 'stop' and 'timecont'.

however, we did not have any keys for 'taskcont'.  the script to do this
exists, and in practce we found that, unlike our initial conclusion that
we would not need this, we have in fact found several occasions where we
want to use it.

so this commit reclaims the 's' key to conform to the task/time
capital/lowercase paradigm described above (which then provides both a
taskcont and timecont), move M-T bind 'tasktmp' to M-X which already had
M-x being 'timetmp' (so now we follow capital/lower paradigm), which
frees up M-T to act as 'taskstop'.  as mentioned above there is no
timestop needed, and we already have M-t used to display taskline, so
this all fits better.