+ jcsrc: keys: M-R: change from redisplay to rptaskallmenu
we have modified rptaskmenu to pass -o/--open to timerecent, which shows
only tasks that have not been completed.  but we added another
invocation, 'rptaskallmenu', that shows all of them and this adds that
bind

'redisplay' is indeed a useful command to have, but only in a few
scenarios and we don't use it often, whereas we use the task menu quite
often and want both ways available (with and without completed tasks).
if it proves in practice that we never use the 'all' variant, we can
switch this back to map to rp 'refresh'