+ jcsrc: keys: M-I: like M-t, but use time at current physical location
we use Pacfic for all our computers just because we always have (since
original move there back around turn of century).  but this leads to
frequently starting a shell to run 'date' with the current time zone (or
looking at our phone)

we'll have to maintain it as we change locations, but that's ok, it's
not that frequent.

we probably should have a UTC time key also, but that's used less
frequently.  maybe we should not use Pacific anymore but rather have the
computer use UTC? now there's an idea...